+++ src/main/java/de/guntram/mcmod/durabilityviewer/config/ModMenuImpl.java
@@ -2,15 +2,10 @@ package de.guntram.mcmod.durabilityviewer.config;
 
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
-import de.tobi.voxelconfig.ConfigScreen;
-import net.minecraft.network.chat.Component;
+import de.guntram.mcmod.durabilityviewer.DurabilityViewer;
 
 public class ModMenuImpl implements ModMenuApi {
     public ConfigScreenFactory<?> getModConfigScreenFactory() {
-        return (screen) -> ConfigScreen.create(
-                Component.translatable("durabilityviewer.gui.title.configs"),
-                Configs.INSTANCE,
-                screen,
-                Configs::saveToFile);
+        return DurabilityViewer::createConfigScreen;
     }
 }